fixes: #11921 - AimsVoronoi: add support for other data types than S16.
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@65576 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsalgo/src/AimsGraphTransferLabels/graphtransferlabels.cc
@@ -239,7 +239,7 @@ int main( int argc, const char** argv )
 
       // voronoi
       cout << "computing voronoi map..." << endl;
-      voro = AimsVoronoiFrontPropagation( voro, -10, -11, 3, 3, 3, 50 );
+      voro = AimsVoronoiFrontPropagation( voro, (short)-10, (short)-11, 3, 3, 3, 50 );
 
       if( !wv.fileName().empty() )
         wv.write( voro );
